+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added URDF and XML writers.
+* Added `compas.robots.RobotModel.to_urdf_file`.
+* Added `compas.files.URDF.from_robot`.
+
 ### Changed
+* Fixed default value for `compas.robots.Axis`.
 
 * Changed surface to mesh conversion to include cleanup and filter functions, and use the outer loop of all brep faces.
 
@@ -57,7 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed bug in `__getstate__`, `__setstate__` of `compas.base.Base`.
 * Fixed bug in `compas_rhino.artists.MeshArtist` and `compas_rhino.artists.NetworkArtist`.
 * Changed length and force constraints of DR to optional parameters.
-* Removed `ABCMeta` from the list of base clases of several objects in compas.
+* Removed `ABCMeta` from the list of base classes of several objects in compas.
 
 ### Removed